[Tab Content Manager] Fix Static Layer Cache
Presently, the static layer cache grows unboundedly. This is a result of
the layers in the cache getting created anew for every content layer.

If a thumbnail for the ThumbnailLayer was ever evicted, the Layer
could still exist, but the texture would be lost leading to both an
empty layer existing and the layer never picking up any changes to
Thumbnails. Moreover, attempting to "load" thumbnails for a
ThumbnailLayer did the work of creating the Thumbnail, but never linked
it to the ThumbnailLayer resulting in wasted effort.

Shift the paradigm entirely such that it is necessary to call
UpdateVisibleTabIds before requesting a static layer. Now each
ThumbnailLayer will have an entry in the ThumbnailCache or will soon
(which TabContentManager will listen for and update accordingly).

This should result in a hard limit on the number of layers and should
ensure there are no ThumbnailLayers without Thumbnails.

Bug: 1402843
Change-Id: I68070f24adee6dce16e1c6a6464ffcabee2f3b50
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4250717
Reviewed-by: David Trainor <dtrainor@chromium.org>
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1128180}