[FreezeCast] Ensure Adding MMCH successfully calls GetMediaController
This ensures that the internal routes observer is updated and
SetRotuesForProvider is called before trying to call
AddMirroringMediaControllerHost, since this function actually relies
on those values being set. We should not NotifyObservers before
AddMirroringMediaControllerHost, since some observers use
OnRoutesUpdated to know to query MediaRouter for
MirroringMediaControllerHosts, hence the original ordering.

This change adds back a test which was accidentally deleted, and
modifies it to ensure that GetMediaController is successfully called
when adding a MirroringMediaControllerHost.

Bug: b:277358204
Change-Id: Ib8009e6de07ddeb012a8b7edced25eb10d5911a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4409582
Reviewed-by: Takumi Fujimoto <takumif@chromium.org>
Commit-Queue: Benjamin Zielinski <bzielinski@google.com>
Cr-Commit-Position: refs/heads/main@{#1128187}