cellular: Update eSIM install retry mechanism
This change updates the retry mechanism to no longer use an exponential
backoff, and instead we now simply wait an entire day before retrying.
This is the first, smaller change to vastly reduce the number of retries
that we see with more improvements to keep track of the number of
failures through reboots.

The on-DUT verification was done by checking that we attempt to install
an eSIM profile once on startup.

Bug: b/266743564, b/276331006
Test: Updated unit tests and verified on DUT.
Change-Id: Ie9e9a69bf30f604b26b7e59e8c160c768f47eb15
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4390524
Commit-Queue: Chad Duffin <chadduffin@chromium.org>
Reviewed-by: Jason Zhang <jiajunz@google.com>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1128193}