Add missing function call to stop flickering when restoring window.
As part of https://crrev.com/c/3714393, HandleWindowMinimizedOrRestored
was overridden in BrowserDesktopWindowTreeHostWin, and the parent
function in BrowserDesktopWindowTreeHostWin is not called. This caused
flickering of vertical bars along the entire screen when restoring the
window. This change adds a call to the parent function.

Bug: 1418895
Change-Id: Icee89e2052c64168f570cf7dff4ad7b499214fa7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4385707
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Reviewed-by: Bruce Dawson <brucedawson@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1128198}