[Pcct-SideSheet] Ensure the icon positioning is more predictable
This CL moves the maximization button into the frame layout used for
location bar. The positioning/visibility of the button is more
predictable as its position is always in the right order against the
other toolbar elements. If there's not enough room, the button gets
hidden.

Bug: 1429476, 1428799
Change-Id: I11d3f84f2eadb145c2b24c215fc7c7ba726a18c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4402512
Reviewed-by: Jinsuk Kim <jinsukkim@chromium.org>
Reviewed-by: Sinan Sahin <sinansahin@google.com>
Commit-Queue: Jinsuk Kim <jinsukkim@chromium.org>
Reviewed-by: Kevin Grosu <kgrosu@google.com>
Cr-Commit-Position: refs/heads/main@{#1128200}