Make default constructors for Encryptor private.
Previously to crbug.com/1427202 being fixed, default
constructors for classes with mojo typemaps that were
to be placed in containers had to be public.

Now this issue is solved, Encryptor::Key no longer needs
to have a default constructor and the default constructor
for Encryptor can be made private as it is only used
by the friended OSCryptAsync factory class.

The public constructors are now protected by a
mojo::DefaultConstruct::Tag passkey pattern.

BUG=1373092

Change-Id: I7cfd2fb5b4ce25638981177b3700b971394bf52d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4410522
Reviewed-by: Alex Gough <ajgo@chromium.org>
Commit-Queue: Will Harris <wfh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1128204}