[favicon] Add a default favicon getter that accounts for bg color
Currently favicon::GetDefaultFavicon() leverages the global native
theme instance to determine whether to serve the light or dark
favicon.

This CL adds a GetDefaultFaviconModel() that takes into account the
color of the background. This enables the correct favicon to be
served for a given context (background colors may be light/dark
regardless of the global color scheme setting).

Bug: 1431468
Change-Id: Iac39f9a81ad5fa8fe1e78885662843e7cf459b35
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4402157
Commit-Queue: Thomas Lukaszewicz <tluk@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1128205}