Align EVRootCAMetadata on Windows and non-Windows
EVRootCAMetadata used the dotted ASCII form of OIDs on Windows and the
DER representation on non-Windows because, historically, that was more
convenient on CertVerifyProcWin. As a result there are a pile of ifdefs
and abstractions like WithBytes to smooth this over.

We probably could have aligned this earlier but now that
CertVerifyProcWin is gone, we especially don't need this. Tidy this up.
While I'm here, remove an unnecessary StringPiece-to-string conversion
by using heteregenous lookup in std::set (https://abseil.io/tips/144).

Bug: 1380536
Change-Id: Icb0e4baa7a0c63dad8dd85c3f00e842f2b2c6941
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4404342
Reviewed-by: Matt Mueller <mattm@chromium.org>
Commit-Queue: David Benjamin <davidben@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1128208}