Restrict OmniboxActions offerings based on AutocompleteMatch position
This change ensures that only specific type of an OmniboxAction is
offered for AutocompleteMatches placed at specific positions on the
list:
- Position 0 and 1: all OmniboxActionIds
- Position 2: HISTORY_CLUSTERS and PEDALs,
- Position 3 and beyond: HISTORY_CLUSTERS only.

The change guarantees that the different kinds of OmniboxAction are
not mixed together - only one type of OmniboxAction can remain
attached to the suggestion. The allowed types are supplied in a form
of a preference list; if no allowed types are found, all Actions are
removed.

Bug: 1418077
Change-Id: Ie56bbc697be0f97361ee081e3001af1a38514a3b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4402397
Reviewed-by: Patrick Noland <pnoland@chromium.org>
Reviewed-by: Orin Jaworski <orinj@chromium.org>
Commit-Queue: Tomasz Wiszkowski <ender@google.com>
Cr-Commit-Position: refs/heads/main@{#1128230}