[CrOS Hotspot] Add the 'Turn Wifi On' button to hotspot notification
A notification is displayed to the user when they turn on the hotspot
while WiFi is enabled. This notification provides a button to the user
to turn the WiFi on which will disable the hotspot as well.

Screenshot: https://screenshot.googleplex.com/5zxNrqngmcez3dj
Bug: b/269353987
Change-Id: Ic45e4d6687d47c15905859885dee698c0d1136d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4408388
Reviewed-by: Jason Zhang <jiajunz@google.com>
Commit-Queue: Nikhil Nayunigari <nikhilcn@google.com>
Cr-Commit-Position: refs/heads/main@{#1128234}