Task manager (Cocoa) - let Enter/return kill processes
The Windows/Views version of the task manager kills the
selected processes when Enter is pressed. This will bring
the macOS/Cocoa version to feature parity in this regard,
while the View version is still in progress.

setKeyEquivalent:@"\r" on an NSButton works two-fold -
- It makes the button respond to Enter/return.
- It marks the button as the active button, visually.
Added that and a browser test to verify the tab is killed.

There are still some differences -
- There is no keyboard shortcut for bringing up the
  task manager.
- The active tab is not always automatically selected
  when the task manager is brought up.
- Ending a process does not select the next entry,
  so it is not possible to click "End process" or press
  Enter/return multiple times to end the next tasks.


Change-Id: I71b2deed3f23949b8d1b1ba675048398855eb368
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4373944
Commit-Queue: PhistucK Productions <phistuck@gmail.com>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1128287}