media: Add ClearKey to Android to support content_browsertests
This CL first adds the key mapping for the clear key UUID to register
as being supported.

Then on all Android modes (e.g Webview), we add a ClearKey MediaBridge
delegate so that we can choose which UUID to return based on which key
system is given.

This CL also modifies some code in the MediaDrmBridge* to not create
storage, as ClearKey does not need storage, and to return a DrmBridge
that corresponds to ClearKey.

Since we ran into some issues with the Android implementation of CK,
this CL does not fully get content_browsertests working on Android.
Instead, this CL commits the necessary code, and marks with a TODO where
all the code needs to be modified to be able to run the browser tests
on Android. The Android work is tracked on b/271451225

Bug: b/263310318
Change-Id: Ic2867b15988eac2e0b3873e705fa9f01d75f28f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4182271
Reviewed-by: Bo Liu <boliu@chromium.org>
Commit-Queue: Vikram Pasupathy <vpasupathy@chromium.org>
Reviewed-by: John Rummell <jrummell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1128319}