Add cookie name or value non-ASCII metrics
These metrics check for non-ASCII values in a cookie's name or value
field. Informally, this metric is interested in non-ASCII Unicode
characters within a cookie's name or value.

These metrics are recorded whenever a cookie is created via
CanonicalCookie::Create(), regardless whether or not the cookie is
ultimately saved to the cookie store.

Bug: 1378154
Change-Id: I46160e81e761792c67c13280ca4b4930a7add847
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4412115
Commit-Queue: Steven Bingler <bingler@chromium.org>
Reviewed-by: David Benjamin <davidben@chromium.org>
Reviewed-by: Chris Fredrickson <cfredric@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1128347}