[color] Update bookmarks to get the favicon using ColorProvider
This CL updates the GetDefaultFavicon() getter used in various
bookmarks classes to use GetDefaultFaviconModel(). This allows
rasterizing the favicon using a context-specific ColorProvider.

Previously this was using a getter that looked only at global
dark/light color scheme, which may not necessarily be correct for
the context.

Bug: 1431468
Change-Id: I470318a7523644bbc52fe528bd52ba30cf15d541
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4411877
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Thomas Lukaszewicz <tluk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1128349}