Fix crashes happens when closing by Widget::CloseNow, delete window.
This is reland of crrev.com/c/4407882 with leak fix. (in PS2)

The `surface_destroyed_callback_` is used to delete the
shell when a client deletes the root surface, so don't call it
if the shell is being destroyed by server side.

Reset delegate properly and don't call twice.

I also added a test case for crbug.com/979128

Bug: b/276351837, crbug.com/979128
Test: covered by unittests

Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4407882
Commit-Queue: Mitsuru Oshima <oshima@chromium.org>
Reviewed-by: Eliot Courtney <edcourtney@chromium.org>
Cr-Original-Commit-Position: refs/heads/main@{#1128092}
Change-Id: I055c54cbc1f76c678850d4645aae43d1c05527d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4406853
Reviewed-by: Yuzhu Shen <yzshen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1128357}