Move CRLSet updates to the CertVerifierWithUpdatableProc path
This avoids an unnecessary IPC round trip to the network service and back, and avoids needing to keep an unparsed copy of the CRLSet data in memory. The network service doesn't need to know about CRLSets and was just acting as an unnecessary intermediary to pass them back to the cert verifier through the old SetConfig path.

Not flag guarded as this has high test confidence and would be very messy to support both paths. See CRLSetBrowserTest.*, TrialComparisonCertVerifierFeatureEnabledTest.TrialEnabledPrefEnabledCRLSetUpdate, CRLSetComponentInstallerTest.*, CertVerifierServiceFactoryCRLSetTest.*.

Bug: 1046728
Change-Id: Ie87383d554afbc1c7299df91ce838d57ebb17901
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4375983
Commit-Queue: Matt Mueller <mattm@chromium.org>
Reviewed-by: Sorin Jianu <sorin@chromium.org>
Reviewed-by: Adam Rice <ricea@chromium.org>
Reviewed-by: Matthew Denton <mpdenton@chromium.org>
Reviewed-by: Mustafa Emre Acer <meacer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1128360}