Remove the 1024 pixel limit to WebCursor bitmaps
WebCursors are already limited to 128 DIP on the renderer, and 150 DIP
on the browser. Checking the underlying bitmap for a max size is
redundant and unneeded.

For the same reasons, remove the upper bound to the scale factor.

Bug: 1149906
Change-Id: Ic819fadbb72f5e0fd920e9ba36466332704010ff
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4410792
Commit-Queue: Henrique Ferreiro <hferreiro@igalia.com>
Commit-Queue: Mike Wasserman <msw@chromium.org>
Reviewed-by: Mike Wasserman <msw@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1128384}