Clean up touch selection menu anchor test.
1. Delete constants and logic in the test file that rely on constants in
   TouchSelectionMenuViews and replace it with a GetMenuWidth test api.

2. Rewrite some comments to make the test clearer.

3. Move tests into an anonymous namespace.

Probably want to add a separate test to check that the menu width is
adjusted when different commands are enabled, especially since deleting
the constants means we don't need separate cases for the anchor rect
adjusting test. Will try to add in a followup.

Bug: b:264825317
Change-Id: Ic4e5f7aceb544f61a75e1fb287f93c899ec87f10
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4406372
Reviewed-by: Thomas Lukaszewicz <tluk@chromium.org>
Commit-Queue: Michelle Chen <michellegc@google.com>
Cr-Commit-Position: refs/heads/main@{#1128432}