AddressTrackerLinux generates diffs
AddressTrackerLinux maintains an AddressMap and a set of online links.
It receives diffs from the kernel that it applies to its AddressMap
and set of online links. This data will also need to be synced to the
network service which won't run AddressTrackerLinux but will still need
to maintain an AddressMap and set of online links.

So add a SetDiffCallback() to AddressTrackerLinux, which will call the
provided callback with diffs whenever the kernel notifies it. In a
future CL these diffs will be sent to the network service.

Bug: 1312226, 1383352
Change-Id: Ie8aeb18e14c1aa08daba1bf0d4dd0443e97ec3d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4407939
Commit-Queue: Matthew Denton <mpdenton@chromium.org>
Reviewed-by: Adam Rice <ricea@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1128567}