Remove unused QueryTileFakeServer class
This CL removes an unused file. QueryTileFakeServer was added in
crrev.com/c/2218683. The only class with an inbound dependency on
it was OmniboxQueryTileSuggestionTest, which was removed in
crrev.com/c/3292366 when the query tile moved to launching on the
NTP rather than omnibox.

Basic file cleanup, no functional changes.


AX-Relnotes: N/A
Bug: N/A
Change-Id: I7444646e2645afe2301f15f4f14022d7798ab858
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4411843
Reviewed-by: Shakti Sahu <shaktisahu@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Commit-Queue: Mark Schillaci <mschillaci@google.com>
Cr-Commit-Position: refs/heads/main@{#1128635}