Add FontSizeAdjust data structure
This is the first patch to bring the two-value syntax of
font-size-adjust into blink. In CSS Font module level 4, the
font-size-adjust CSS property could have a numeric value only,
so we implemented it with a float value. However, the latest standard
draft allows font-size-adjust to have additional info, namely a font
metric along with an adjusting value. We need a placeholder to keep
that information.

This patch does not change the existing behavior of the legacy
font-size-adjust implementation. Instead, it aims to replace the float
placeholder with the FontSizeAdjust data structure.

Bug: 1219875
Change-Id: I99f19540bf44c753436bfc6d4959b527c68d54bd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4378170
Reviewed-by: Dominik Röttsches <drott@chromium.org>
Commit-Queue: Dominik Röttsches <drott@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1128649}