Switch CECPQ2 for Kyber.
This change reuses the plumbing for CECPQ2 and repurposes it for Kyber.
CECPQ2 was an experimental post-quantum KEM and it will be replaced by
NIST's choice, Kyber.

This change also deprecates the CECPQ2Enabled enterprise policy. This
policy allowed enterprises to disable (but not force-enable) CECPQ2.
However, CECPQ2 is disabled by default so it's been moot for a while.

Change-Id: Id3b705e5b691ed8e260b9472b26e242a66397db2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4387832
Reviewed-by: Bob Beck <bbe@google.com>
Reviewed-by: Igor <igorcov@chromium.org>
Reviewed-by: David Benjamin <davidben@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Commit-Queue: Adam Langley <agl@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1128650}