[CodeHealth] Use a unique id to identify CaptionBubbleModels
This adds a unique id to CaptionBubbleModel to identify each instance,
rather than using its raw pointer. This allows removal of a dangling
pointer whose purpose was to identify the active model at the time of
posting a task. The unique id accomplishes the same purpose, and avoids
bugs such as if the model is deallocated and a different model is
created at the same address.

Bug: 1424420
Change-Id: I9c501e2eb772244b0e7d0b1bfc1ddbdf6791cd8b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4378358
Commit-Queue: Lily Chen <chlily@chromium.org>
Reviewed-by: Abigail Klein <abigailbklein@google.com>
Cr-Commit-Position: refs/heads/main@{#1128659}