[iOS] UIButtonConfiguration refactor for manual accessory VC
The password key logo was a different color due to
UIButtonConfiguration changes. This was because the image rendering
mode was set to auto and messed up property functionality when
refactoring a UIButton to a UIButtonConfiguration system. This CL fixes
that and reflect the appropriate password key color and color changes.

Bug: 1432213
Change-Id: Ifb87380c490947da05b1c16314fd73079a229ff2
Fixed: 1424108
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4415236
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Joemer Ramos <joemerramos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1128719}