[dPWA] Do not pass profile to async calls in OS integration
This CL is a pure refactor with no change in logic. Instead of passing
in a profile to async callbacks and tasks posted on other threads, the
profile path is directly passes instead (since that is the only field
that is mostly used). This prevents callbacks from being called on
profile shutdown, and decouples the lifecycle of the profile from
OS integration logic.

Bug: 1425857, b/274466685
Change-Id: Icc35239252df4c145276d3d193cce8513104aa05
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4415199
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Commit-Queue: Dibyajyoti Pal <dibyapal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1128734}