Add unit test for the PassesLuhnCheck function
This patch adds a simple unit test for the PassesLuhnCheck function. It
simply expects a single Luhn-valid number to pass the Luhn check and a
single Luhn-invalid number to fail the Luhn check.

In the future we should probably add more Luhn valid and invalid numbers
in kValidNumbers and kInvalidNumbers, as currently the former only has
*some* Luhn-valid numbers and the latter has only 1 Luhn-invalid number.

Bug: 927767
Change-Id: I8c752e4e123b8b5a95225deb0cd6f59dc5abc033
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4411174
Commit-Queue: Stephen McGruer <smcgruer@chromium.org>
Reviewed-by: Matthias Körber <koerber@google.com>
Reviewed-by: Stephen McGruer <smcgruer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1128739}