[DIPS] Make kBounceThresholdSeconds a DIPS' feature parameter
This CL is the first to a series of changes to make this value
alterable via field trial parameterization.

The current changes make this value a feature param for the DIPS
feature.

A foreseeable next step will be to push the necessary field trial
configs (initially via fieldtrial_testing_config.json) with this param
included.

Bug: 1431309
Change-Id: I065d9c1ea68fcd5b54280239fa32401f95d55f7f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4414632
Reviewed-by: Ben Kelly <wanderview@chromium.org>
Commit-Queue: Jonathan Njeunje <njeunje@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1128783}