[Start] Build histograms for click rates of Profile button on Start and NTP
We added a new metric to the two histograms (StartSurface.Module.Click and NewTabPage.Module.Click) in this CL. This metric keeps track of the number of times a user clicks on the profile button on Start/NTP. The histogram StartSurface.Module.Click is updated with label "Profile button" each time a user clicks on the profile button on Start surface. The histogram NewTabPage.Module.Click is updated with label "Profile button" each time a user clicks on the profile button on New tab page.

Bug: 1384156
Change-Id: Ib7ece948c323d76cdd0a48e6e24616357b85e184
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4394362
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Reviewed-by: Xi Han <hanxi@chromium.org>
Reviewed-by: Weilun Shi <sweilun@chromium.org>
Commit-Queue: Xinyi Ji <xinyiji@chromium.org>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1128798}