Add CUJs to verify launching an app opens it in the right profiles.
On Mac OS launching from platform shortcut should open the app in all
the profile it was open in when the app was last quit. These CUJs
verify this behavior.

I broke this logic in https://crrev.com/c/4013413, so this CL also
includes a fix for that.

Finally, WebAppIntegrationTestDriver::LaunchFromPlatformShortcut
is fixed in this CL to properly wait for all launches triggered
to be finished. SetMacShimStartupDoneCallbackForTesting was added
for such purposes previously, although only used for file handling
launches. There were a couple of cases where the shim-startup-done
callback wasn't correctly getting called, so this CL also fixes those.

Bug: 1427296
Change-Id: Ieb22eb16d72d56027b867106c7027346cb4f0e95
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4410410
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Commit-Queue: Marijn Kruisselbrink <mek@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1128834}