update_rust.py will try the latest and fallback revisions
Have --print-package-version print the stamp version, as the Clang
update.py script does too. This tells the actual version of the
toolchain in the repo.

When updating Rust, check if the latest Rust/Clang revisions have a
Rust toolchain package. Use if it so. If not, try the
FALLBACK_REVISION.

After we have chosen a version, then proceed to check if the stamp
matches, and download a replacement toolchain if not.

This should mean that clang rolls no longer need to bump the
FALLBACK_REVISION to succeed. We can do it after the fact if we
find that we can't roll Rust, to pin it to something newish.

Bug: 1401042
Change-Id: Ic0436e87d18566f3498d30f3560fcc4e196ec240
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4406712
Reviewed-by: Collin Baker <collinbaker@chromium.org>
Commit-Queue: danakj <danakj@chromium.org>
Reviewed-by: Zequan Wu <zequanwu@google.com>
Cr-Commit-Position: refs/heads/main@{#1128845}