Fix web app sync integration test CUJs on Mac.
Fix the CUJs for these tests to have the correct expectations on Mac.
Also re-enable one test that was accidentally left disabled previously,
and re-enable another test that was fixed by crrev.com/c/4410410.

Also tweaks the Mac LaunchFromPlatformShortcut implementation a bit
to better support the case where a launch causes a app window in a
different profile than the "current" on to be activated.

Bug: 1427296
Change-Id: Ia8be974a2e560ab0dfded73fb92a333f12d8a55a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4415000
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Commit-Queue: Marijn Kruisselbrink <mek@chromium.org>
Reviewed-by: Rushan Suleymanov <rushans@google.com>
Cr-Commit-Position: refs/heads/main@{#1128852}