Smooth touch drag start animation
When doing a touch drag, DragDropController scales and shifts the drag
image to 120% with 25dp vertical offset. This is intended as a visual
cue for the drag operation.

AppListItemView handles its own drag cues by only scaling up the app
item by 120% without shifting the cursor anchor.

This CL tries to compensate for the operations that DragDropController
attempts by returning the icon to the origina scale before invoking
RunShellDrag. It also attempts to compensate for the anchor shift by
invoking WriteDragData with a different anchor position.

While not completely smooth, some more polishing may need to be
revised to improve the look of the animation.

Bug: b:261985897, b:277657912
Change-Id: I5ff12cdcb21fc64f2a726e8b5563e75cf5206f2a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4412820
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Commit-Queue: Ana Salazar <anasalazar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1128855}