Put build scripts in a better place
The .exe already has the target name in it, so it doesn't need to add
that to the directory path.

The OUT_DIR env var should point to gen/ not to out/, so we point it
to $target_gen_dir/$target_name, so that it is unique for multiple
GN targets in the same BUILD.gn file (and thus same target_gen_dir).
Often the OUT_DIR is used by build scripts, but it can be used
without one too, and is chosen in rust_target() which doesn't know if
a build script exists. So we don't tie it to the build script itself.

R=collinbaker@chromium.org

Bug: 1271215
Change-Id: I4e78e7656d7bdb99a91f7fe42d2d35fd7819a7eb
Cq-Include-Trybots: luci.chromium.try:android-rust-arm32-rel,android-rust-arm64-dbg,android-rust-arm64-rel,linux-rust-x64-dbg,linux-rust-x64-rel
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4401332
Reviewed-by: Collin Baker <collinbaker@chromium.org>
Commit-Queue: danakj <danakj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1128856}