Only identify new alterable extensions when we can show them.
ExtensionErrorController::ShowErrorIfNeeded might be triggered
multiple times back to back while the notification is opened. However.
`IdentifyAlertableExtensions` may not always return same extensions,
especially for extensions that are blocked by policies. As when they
are moved from enabled list to disabled list due to the policy, we
will no longer include them in the error UI.

If IdentifyAlertableExtensions reset the list of blocked extensions,
they won't be marked as "acknowledged" properly after notification is
confirmed. Hence the notification may be opened for it more than once.

The fix here will only call IdentifyAlertableExtensions to recalculate
the blocklist when there is no on-going error dialog.


Change-Id: Ibde9cdc5d73b92e96c2d33604a48665f349a9a93
Bug: 1428646
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4401017
Commit-Queue: Owen Min <zmin@chromium.org>
Reviewed-by: David Bertoni <dbertoni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1128868}