[Extensions] Use pref value for file access on installed extension load
This CL modifies the logic when loading an installed extension to base
the file access purely on the state of the related pref, rather than
any previously stored creation flags. Normally these should be aligned,
but in case they are not we want to stick to what the pref says as that
is the state we display to the user on the extension management page.

Also expands a related test and adds some more explicit testing for this
mismatch case.

Bug: 1385343, 1432284
Change-Id: I7c45a7a174665ee138c4a7a3fba77b3bbe381508
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4313067
Commit-Queue: Tim <tjudkins@chromium.org>
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1128884}