ozone: Implement GetSupportedFormatsForTexturing for Ozone/Wayland.
This CL implements the
SurfaceFactoryOzone::GetSupportedFormatsForTexturing() method for the
Ozone Wayland backend. This is necessary because when doing software
video decoding, capabilities.image_ycbcr_420v [1] is determined by
whether the returned formats contain gfx::BufferFormat::YUV_420_BIPLANAR
[2], and this helps determine if it's possible to copy software decoded
frames into CPU-mappable native GpuMemoryBuffers which is necessary to
promote such frames to hardware overlays.

The implementation in this CL follows the existing implementation for
the Ozone DRM backend which uses the gbm API to query the formats that
the gbm implementation can allocate for texturing [3].

In this CL, instead of opening the render node for creating a temporary
GBM device, we reuse the existing GBM device tracked by the
WaylandBufferManagerGpu. Note that we obtain this device by calling
WaylandBufferManagerGpu::GetGbmDevice() which needs
gl::GLSurfaceEGL::GetGLDisplayEGL()->ext->b_EGL_EXT_image_dma_buf_import
to be true [4]. Because of that, I had to move the
GetSupportedFormatsForTexturing() call to later in the process startup
after the call to gl::init::InitializeExtensionSettingsOneOffPlatform()
by which point, I assume, the status of EGL_EXT_image_dma_buf_import is
known. There is a note about needing to call
GetSupportedFormatsForTexturing() before sandboxing in order to not
break camera preview. This CL doesn't actually change that: the call to
EnsureSandboxInitialized() still comes later except for the case in
which gpu_sandbox_start_early (and in that case the call to
GetSupportedFormatsForTexturing() was after that even before this CL).

[1] https://source.chromium.org/chromium/chromium/src/+/main:content/renderer/media/gpu/gpu_video_accelerator_factories_impl.cc;l=366;drc=694cf36503b29e1bcdc2ab3441e9379ded0137b4
[2] https://source.chromium.org/chromium/chromium/src/+/main:gpu/command_buffer/service/feature_info.cc;l=198-200;drc=3389640bcb39f116ba7307265f9f52ef099b3bb0
[3] https://source.chromium.org/chromium/chromium/src/+/refs/heads/main:ui/ozone/platform/drm/gpu/gbm_surface_factory.cc;l=201-240;drc=6315c6806f72d3437853a69721dbc7659fe3fe1e
[4] https://source.chromium.org/chromium/chromium/src/+/refs/heads/main:ui/ozone/platform/wayland/gpu/wayland_buffer_manager_gpu.cc;l=323-329;drc=39c98010d31f70924bec7a282b8f027b171b1329

Bug: b:277387200
Test: software video playback in lacros-chrome on a volteer device
Change-Id: I684331a6c537e57e01be95a8e9adfde36f1e72b5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4410380
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Reviewed-by: Kramer Ge <fangzhoug@chromium.org>
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Maksim Sisov <msisov@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1128931}